requirements.txt
absl-py==2.1.0
accelerate==0.32.1
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
av==10.0.0
bitsandbytes==0.42.0
Brotli==1.0.9
cachetools==5.3.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.1
ctranslate2==4.3.1
cycler==0.12.1
datasets==2.18.0
decorator==4.4.2
dill==0.3.7
dnspython==2.6.1
dpcpp-cpp-rt==2024.2.0
evaluate==0.4.2
exceptiongroup==1.2.2
facenet-pytorch==2.5.3
fastapi==0.109.2
ffmpy==0.3.1
filelock==3.12.3
flatbuffers==23.5.26
fonttools==4.48.1
frozenlist==1.4.1
fsspec==2023.10.0
future==1.0.0
gmpy2==2.1.5
google-auth==2.27.0
google-auth-oauthlib==1.2.0
gradio==4.17.0
gradio_client==0.9.0
greenlet==3.0.3
grpcio==1.60.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.3
humanfriendly==10.0
idna==3.4
imageio==2.33.1
imageio-ffmpeg==0.4.9
importlib-resources==6.1.1
intel-cmplr-lib-rt==2024.2.0
intel-cmplr-lib-ur==2024.2.0
intel-cmplr-lic-rt==2024.2.0
intel-opencl-rt==2024.2.0
intel-openmp==2024.2.0
intel-sycl-rt==2024.2.0
Jinja2==3.1.2
jiwer==3.0.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.3
librosa==0.10.2.post1
llvmlite==0.42.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.9.1
mdurl==0.1.2
mkl==2024.2.0
mkl-fft==1.3.6
mkl-random==1.2.2
mkl-service==2.4.0
moviepy==1.0.3
mpmath==1.2.1
msgpack==1.0.7
multidict==6.0.5
multiprocess==0.70.15
networkx==3.0
numba==0.59.0
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnxruntime==1.17.0
opencv-python==4.7.0.72
orjson==3.9.13
packaging==23.2
pandas==2.2.0
Pillow==9.4.0
platformdirs==4.2.0
pooch==1.8.0
proglog==0.1.10
protobuf==4.23.4
psutil==5.9.8
pyarrow==15.0.0
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pydub==0.25.1
Pygments==2.17.2
pymongo==4.6.3
pyparsing==3.1.1
pyreadline3==3.4.1
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.8
python-multipart==0.0.7
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.6.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
responses==0.18.0
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
ruff==0.2.1
safetensors==0.4.2
scikit-learn==1.4.0
scipy==1.12.0
seaborn==0.12.2
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
soundfile==0.12.1
soxr==0.3.7
SQLAlchemy==2.0.27
starlette==0.36.3
sympy==1.12
tbb==2021.13.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tokenizers==0.15.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.2.0
torchvision==0.17.0
tqdm==4.66.1
transformers==4.37.2
triton==2.2.0
typer==0.9.0
typing_extensions==4.12.2
tzdata==2023.4
urllib3==2.1.0
uvicorn==0.27.0.post1
websockets==11.0.3
Werkzeug==3.0.1
win-inet-pton==1.1.0
xxhash==3.4.1
yarl==1.9.4
